+++ packages/actors-monitor/__tests__/http-source.test.ts
@@ -42,6 +42,7 @@ const socketsSection = {
     deliveries: 1200,
     deliveryBytes: 48_000,
     throttleQuantized: 6,
+    subscriptionsShed: 0,
     open: 3,
     inFlight: 1,
     subscriptions: 4,